+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.4.1
+
+- Contract Description File add `buildType`
+- remove ``debugUri`` command option, will print **Launch Debugger** uri default
+
 ## 0.3.10
 
 - Fix loc undefined